refactor(core): update core modules for frontend-backend separation
- Update API client for new architecture
- Improve cache management
- Update geocoding engine
- Enhance export functionality
- Update pipeline orchestration
- Add rate limiting improvements
- Update logging configuration

Breaking changes:
- Reorganized module structure
- Updated API client interfaces
- Changed cache management approach
