+++ socialmapper/census/infrastructure/memory.py
@@ -26,7 +26,7 @@ import pandas as pd
 import psutil
 
 from ...config.optimization import MemoryConfig
-from ...ui.console import get_logger
+from ...console import get_logger
 
 logger = get_logger(__name__)
