+++ socialmapper/census/infrastructure/streaming.py
@@ -55,7 +55,7 @@ except ImportError:
     warnings.warn("Polars not available, using pandas for data processing", stacklevel=2)
 
 
-from ...ui.console import get_logger
+from ...console import get_logger
 
 logger = get_logger(__name__)
