+++ socialmapper/distance/__init__.py
@@ -16,7 +16,7 @@ from shapely.geometry import Point
 
 from socialmapper.progress import get_progress_bar
 
-from ..ui.console import get_logger
+from ..console import get_logger
 
 # Import the high-performance engine
 from .engine import ParallelDistanceProcessor, VectorizedDistanceEngine, benchmark_distance_engines
