+++ socialmapper/export/formats/parquet.py
@@ -10,7 +10,7 @@ import pandas as pd
 
 from socialmapper.constants import CATEGORICAL_CONVERSION_THRESHOLD
 
-from ...ui.console import get_logger
+from ...console import get_logger
 from ..base import BaseExporter, ExportError
 
 logger = get_logger(__name__)
