+++ socialmapper/export/utils.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import geopandas as gpd
 import pandas as pd
 
-from ..ui.console import get_logger
+from ..console import get_logger
 
 logger = get_logger(__name__)
