+++ socialmapper/geocoding/cache.py
@@ -10,7 +10,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from ..ui.console import get_logger
+from ..console import get_logger
 from .models import AddressInput, GeocodingConfig, GeocodingResult
 
 logger = get_logger(__name__)
