+++ socialmapper/geocoding/providers.py
@@ -13,7 +13,7 @@ from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
 
 from ..neighbors import get_neighbor_manager
-from ..ui.console import get_logger
+from ..console import get_logger
 from .models import AddressInput, AddressProvider, AddressQuality, GeocodingConfig, GeocodingResult
 
 logger = get_logger(__name__)
