+++ socialmapper/isochrone/clustering.py
@@ -24,7 +24,7 @@ from shapely.geometry import Point
 from sklearn.cluster import DBSCAN
 
 # Setup logging
-from ..ui.console import get_logger
+from ..console import get_logger
 from .travel_modes import TravelMode, get_default_speed, get_highway_speeds, get_network_type
 
 logger = get_logger(__name__)
