+++ socialmapper/isochrone/concurrent.py
@@ -25,7 +25,7 @@ import psutil
 
 from ..constants import HIGH_CPU_USAGE_THRESHOLD, HIGH_MEMORY_USAGE_THRESHOLD
 from ..progress import get_progress_bar
-from ..ui.console import get_logger
+from ..console import get_logger
 from .cache import ModernNetworkCache, download_and_cache_network
 from .clustering import (
     OptimizedPOICluster,
