+++ socialmapper/util/invalid_data_tracker.py
@@ -12,7 +12,7 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any
 
-from ..ui.console import get_logger
+from ..console import get_logger
 
 logger = get_logger(__name__)
