+++ socialmapper/util/rate_limiter.py
@@ -11,7 +11,7 @@ from functools import wraps
 import httpx
 
 # Configure logger
-from ..ui.console import get_logger
+from ..console import get_logger
 
 logger = get_logger(__name__)
